edit permission, tests
